analyses/vggish/app/requirements.txt
dataclasses
firebase-admin
Flask 
google-cloud-pubsub
google-cloud-storage
llvmlite==0.32.1
matplotlib
numpy 
resampy 
six 
soundfile 
tensorflow==1.15 
tf_slim 